+++ src/mock/review.ts
@@ -18,16 +18,4 @@ function generateReview() {
   };
 }
 
-export type Review = {
-  id: number;
-  date: string;
-  user: {
-    name: string;
-    avatarUrl: string;
-    isPro: boolean;
-  };
-  comment: string;
-  rating: number;
-}
-
 export { generateReview };
